test: move WebSocket tests into dedicated directory
