+++ .github/workflows/websocket.yml
@@ -1,4 +1,4 @@
-name: Build and Test
+name: Test and Build WebSocket Layer
 
 on:
   workflow_dispatch:
@@ -79,7 +79,7 @@ jobs:
     - name: Restore uWebSockets source cache
       uses: actions/cache@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
       with:
-        path: tests/interop/uWebSockets
+        path: tests/websocket/interop/uWebSockets
         key: uwebsockets-src-v1-${{ runner.os }}-fe7da4cb05622b8d004718ec3ca05101782eb1c2
 
     - name: Install Python dependencies
@@ -118,15 +118,15 @@ jobs:
 
     - name: Run tests (integration + echo)
       run: |
-        ${{ matrix.os == 'windows-latest' && 'Start-Process -NoNewWindow python -ArgumentList "tests/integration/echo_server.py 9002"' || 'python tests/integration/echo_server.py 9002 &' }}
+        ${{ matrix.os == 'windows-latest' && 'Start-Process -NoNewWindow python -ArgumentList "tests/websocket/integration/echo_server.py 9002"' || 'python tests/websocket/integration/echo_server.py 9002 &' }}
         sleep 3
         ${{ matrix.os == 'windows-latest' && 'build/Release/integration_test.exe' || './build/integration_test' }} ws://127.0.0.1:9002/
         ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }} ws://127.0.0.1:9002/
-        ${{ matrix.asan && 'python tests/perf/long_run.py ws://127.0.0.1:9002/ 0.083 30' || '' }}
+        ${{ matrix.asan && 'python tests/websocket/perf/long_run.py ws://127.0.0.1:9002/ 0.083 30' || '' }}
 
     - name: Stalled TCP test (listen-only server)
       run: |
-        python tests/handshake/listen_only_test.py ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/listen_only_test.py ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
 
     - name: Run unit tests
       run: |
@@ -138,7 +138,7 @@ jobs:
     - name: Build + test uWS echo-server
       env:
         VCPKG_BINARY_SOURCES: clear;files,${{ github.workspace }}/.vcpkg-bincache,readwrite
-      run: python tests/interop/uws_echo_test.py 19004
+      run: python tests/websocket/interop/uws_echo_test.py 19004
 
     - name: Install nginx (Linux)
       if: matrix.os == 'ubuntu-latest'
@@ -158,17 +158,17 @@ jobs:
         echo "C:\nginx-1.27.4" | Out-File -FilePath $env:GITHUB_PATH -Append
 
     - name: Test nginx WebSocket proxy
-      run: python tests/nginx/nginx_proxy_test.py
+      run: python tests/websocket/nginx/nginx_proxy_test.py
 
     - name: Run network I/O tests
-      run: python -m pytest tests/network/test_network.py -v -x
+      run: python -m pytest tests/websocket/network/test_network.py -v -x
 
     - name: Autobahn fuzzingserver (Linux only)
       if: matrix.os == 'ubuntu-latest' && matrix.asan
       run: |
         mkdir -p reports
         docker run -d -p 9001:9001 \
-          -v ${{ github.workspace }}/tests/autobahn:/config \
+          -v ${{ github.workspace }}/tests/websocket/autobahn:/config \
           -v ${{ github.workspace }}/reports:/reports \
           --name autobahn \
           crossbario/autobahn-testsuite \
@@ -252,12 +252,12 @@ jobs:
 
     - name: Run Toxiproxy tests
       run: |
-        python tests/handshake/toxiproxy_test.py slicer pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
-        python tests/handshake/toxiproxy_test.py latency pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
-        python tests/handshake/toxiproxy_test.py timeout pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
-        python tests/handshake/toxiproxy_test.py reset pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
-        python tests/handshake/toxiproxy_test.py bandwidth pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
-        python tests/handshake/bad_handshake_toxiproxy_test.py ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/toxiproxy_test.py slicer pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/toxiproxy_test.py latency pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/toxiproxy_test.py timeout pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/toxiproxy_test.py reset pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/toxiproxy_test.py bandwidth pre ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
+        python tests/websocket/handshake/bad_handshake_toxiproxy_test.py ${{ matrix.os == 'windows-latest' && 'build/Release/echo_test.exe' || './build/echo_test' }}
 
     - name: Stop background servers
       if: always()
