+++ tests/websocket/network/servers/peer_close_server.py
