+++ tests/websocket/network/servers/utf8_fail_fast_server.py
