+++ tests/websocket/unit/test_frame.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "../../src/ws_client.c"
+#include "../../../src/ws_client.c"
 
 static int failed = 0;
